Create timeline charts (xG) using event data.

This function allows to make match timelines using data collected event-by-event.

Usage

plot_timeline(
  data,
  match_year,
  team_home,
  team_away,
  home_color,
  away_color,
  theme = ""
)

Arguments

data

is for the dataset used.

match_year

the year the match was played.

team_home

is for the home team according to data.

team_away

is for the away team according to data.

home_color

is for the colour of the line for the home team.

away_color

is for the colour of the line for the away team.

theme

to select the colours. Choose from 4 themes -> dark, almond, rose, white.

Value

a ggplot2 object

Examples

if (FALSE) {
plot <- plot_timeline(data = data, match_year = 2021, 
                      team_home = "Manchester United", team_away = "Liverpool",
                      home_color = "#e31a1c", away_color = "#980043", theme = "dark")
plot
}