Function for plotting xG Trendline with FBref/ StatsBomb data.

The data can be scraped from FBref.
Dataframe passed in must have the following column names:

Date (format: year-month-day).yyyy-mm-dd,
Home_xG (xG for Home Team),
Away_xG (xG for Away Team),
Home (Home Team),
Away (Away Team)

Usage

plot_trendline(data, team, colour_xg, colour_xga, roll_avg, theme = "")

Arguments

data

is for the dataset used. Select the number of matches wanted in the viz beforehand.

team

is to select the specific team for the viz. Team must be accurate as per FBref specifications.

colour_xg

is for selecting colour for xGoals.

colour_xga

is for selecting the colour for xGoalsAgainst.

roll_avg

is for setting the rolling average for the data.

theme

to select the theme from 4 options -> dark, almond, rose, white.

Details

For best clarity, export plot as a 2000x1000 png

Examples

if (FALSE) {
plot <- plot_trendline(data = pl, team = "Tottenham",
                       colour_xg = "#08519c", colour_xga = "#cb181d",
                       roll_avg = 10, theme = "dark")
plot
}